[refactor] 파일 가져오기 수정
