+++ src/assets/background/background_lobby.png
